Made changes per PR.
