+++ foxpuppet/browser/window.py
@@ -2,14 +2,17 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
+from foxpuppet.windows import Windows
 from selenium.webdriver.common.by import By
 
-from .tabbar import Tabbar
 from .navbar import Navbar
-from foxpuppet.browser.windows import Windows
+from .tabbar import Tabbar
 
 
 class BrowserWindow(object):
+    _file_menu_button_locator = (By.ID, 'file-menu')
+    _file_menu_private_window_locator = (By.ID, 'menu_newPrivateWindow')
+    _file_menu_new_window_button_locator = (By.ID, 'menu_newNavigator')
     _nav_bar_locator = (By.ID, 'nav-bar')
     _tab_browser_locator = (By.ID, 'tabbrowser-tabs')
     _title_bar_close_button_locator = (By.ID, 'titlebar-close')
@@ -50,7 +53,17 @@ class BrowserWindow(object):
 
                 let chromeWindow = arguments[0].ownerDocument.defaultView;
                 return PrivateBrowsingUtils.isWindowPrivate(chromeWindow);
-            """, [self.windows.window_element])
+            """, self.windows.window_element)
+
+    def open_window(self, private=False):
+        self.selenium.set_context('chrome')
+        self.selenium.find_element(*self._file_menu_button_locator).click()
+        if private:
+            self.selenium.find_element(
+                *self._file_menu_private_window_locator).click()
+        else:
+            self.selenium.find_element(
+                *self._file_menu_new_window_button_locator).click()
 
     def close(self):
         self.selenium.close()
