--- foxpuppet/tests/test_open_tab.py
@@ -1,9 +1,0 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-
-class TestTab(object):
-
-    def test_open_new_tab(self):
-        assert 1 == 1
